+++ test/complex.test.ts
@@ -4,20 +4,20 @@ import { ComplexSchema } from "../src/schema";
 describe("Complex type definitions", () => {
     describe("A valid value", () => {
         it("should validate", () => {
-            expect(validateComplex("a", { scalar: "string" })).toBeTruthy();
-            expect(validateComplex(0, { scalar: "number" })).toBeTruthy();
-            expect(validateComplex(true, { scalar: "boolean" })).toBeTruthy();
+            expect(validateComplex("a", { string: {} })).toBeTruthy();
+            expect(validateComplex(0, { number: {} })).toBeTruthy();
+            expect(validateComplex(true, { boolean: {} })).toBeTruthy();
         });
     });
 
     describe("An invalid value", () => {
         it("should not validate", () => {
-            expect(validateComplex(0, { scalar: "string" })).toBeFalsy();
-            expect(validateComplex(true, { scalar: "string" })).toBeFalsy();
-            expect(validateComplex("", { scalar: "number" })).toBeFalsy();
-            expect(validateComplex(false, { scalar: "number" })).toBeFalsy();
-            expect(validateComplex("true", { scalar: "boolean" })).toBeFalsy();
-            expect(validateComplex(1, { scalar: "boolean" })).toBeFalsy();
+            expect(validateComplex(0, { string: {} })).toBeFalsy();
+            expect(validateComplex(true, { string: {} })).toBeFalsy();
+            expect(validateComplex("", { number: {} })).toBeFalsy();
+            expect(validateComplex(false, { number: {} })).toBeFalsy();
+            expect(validateComplex("true", { boolean: {} })).toBeFalsy();
+            expect(validateComplex(1, { boolean: {} })).toBeFalsy();
         });
     });
 
@@ -59,68 +59,67 @@ describe("Complex type definitions", () => {
 
     describe("A valid string", () => {
         it("should validate", () => {
-            expect(validateComplex("", { scalar: "string" })).toBeTruthy();
+            expect(validateComplex("", { string: {} })).toBeTruthy();
         });
     });
 
     describe("An invalid string", () => {
         it("should not validate", () => {
-            expect(validateComplex(5, { scalar: "string" })).toBeFalsy();
+            expect(validateComplex(5, { string: {} })).toBeFalsy();
         });
     });
 
     describe("A valid number", () => {
         it("should validate", () => {
-            expect(validateComplex(0, { scalar: "number" })).toBeTruthy();
+            expect(validateComplex(0, { number: {} })).toBeTruthy();
         });
     });
 
     describe("An invalid number", () => {
         it("should not validate", () => {
-            expect(validateComplex("5", { scalar: "number" })).toBeFalsy();
+            expect(validateComplex("5", { number: {} })).toBeFalsy();
         });
     });
 
     describe("A valid boolean", () => {
         it("should validate", () => {
-            expect(validateComplex(true, { scalar: "boolean" })).toBeTruthy();
+            expect(validateComplex(true, { boolean: {} })).toBeTruthy();
         });
     });
 
     describe("An invalid boolean", () => {
         it("should not validate", () => {
-            expect(validateComplex(1, { scalar: "boolean" })).toBeFalsy();
+            expect(validateComplex(1, { boolean: {} })).toBeFalsy();
         });
     });
 
     describe("A valid nullable value", () => {
         it("should validate", () => {
-            expect(validateComplex(null, { scalar: "boolean", nullable: true })).toBeTruthy();
+            expect(validateComplex(null, { boolean: {}, nullable: true })).toBeTruthy();
         });
     });
 
     describe("An invalid nullable value", () => {
         it("should not validate", () => {
-            expect(validateComplex(null, { scalar: "boolean" })).toBeFalsy();
+            expect(validateComplex(null, { boolean: {} })).toBeFalsy();
         });
     });
 
     describe("A valid optional value", () => {
         it("should validate", () => {
-            expect(validateComplex(undefined, { scalar: "boolean", optional: true })).toBeTruthy();
+            expect(validateComplex(undefined, { boolean: {}, optional: true })).toBeTruthy();
         });
     });
 
     describe("An invalid optional value", () => {
         it("should not validate", () => {
-            expect(validateComplex(undefined, { scalar: "boolean" })).toBeFalsy();
+            expect(validateComplex(undefined, { boolean: {} })).toBeFalsy();
         });
     });
 
     describe("A valid string with a regex matcher", () => {
         it("should validate", () => {
             expect(validateComplex("I am happy", {
-                scalar: "string",
                 string: {
                     matcher: /[ ]am[ ]/
                 }
@@ -131,7 +130,6 @@ describe("Complex type definitions", () => {
     describe("An invalid string with a regex matcher", () => {
         it("should not validate", () => {
             expect(validateComplex("I am happy", {
-                scalar: "string",
                 string: {
                     matcher: /[ ]are[ ]/
                 }
@@ -142,7 +140,6 @@ describe("Complex type definitions", () => {
     describe("A valid string with a function matcher", () => {
         it("should validate", () => {
             expect(validateComplex("I am happy", {
-                scalar: "string",
                 string: {
                     matcher: s => s.includes(" am ")
                 }
@@ -153,7 +150,6 @@ describe("Complex type definitions", () => {
     describe("An invalid string with a function matcher", () => {
         it("should not validate", () => {
             expect(validateComplex("I am happy", {
-                scalar: "string",
                 string: {
                     matcher: s => s.includes(" are ")
                 }
@@ -164,7 +160,6 @@ describe("Complex type definitions", () => {
     describe("A valid string with a value matcher", () => {
         it("should validate", () => {
             expect(validateComplex("I am happy", {
-                scalar: "string",
                 string: {
                     matcher: "I am happy"
                 }
@@ -175,7 +170,6 @@ describe("Complex type definitions", () => {
     describe("An invalid string with a value matcher", () => {
         it("should not validate", () => {
             expect(validateComplex("I am happy", {
-                scalar: "string",
                 string: {
                     matcher: "I are happy"
                 }
@@ -186,7 +180,6 @@ describe("Complex type definitions", () => {
     describe("A valid string without a matcher", () => {
         it("should validate", () => {
             expect(validateComplex("I am happy", {
-                scalar: "string",
                 string: {}
             })).toBeTruthy();
         });
@@ -195,7 +188,6 @@ describe("Complex type definitions", () => {
     describe("A valid number with a specific value", () => {
         it("should validate", () => {
             expect(validateComplex(5, {
-                scalar: "number",
                 number: {
                     value: 5
                 }
@@ -206,7 +198,6 @@ describe("Complex type definitions", () => {
     describe("An invalid number with a specific value", () => {
         it("should not validate", () => {
             expect(validateComplex(4, {
-                scalar: "number",
                 number: {
                     value: 5
                 }
@@ -217,7 +208,6 @@ describe("Complex type definitions", () => {
     describe("A valid number with a specific minimum", () => {
         it("should validate", () => {
             expect(validateComplex(5, {
-                scalar: "number",
                 number: {
                     min: 5
                 }
@@ -228,7 +218,6 @@ describe("Complex type definitions", () => {
     describe("An invalid number with a specific minimum", () => {
         it("should not validate", () => {
             expect(validateComplex(4, {
-                scalar: "number",
                 number: {
                     min: 5
                 }
@@ -239,7 +228,6 @@ describe("Complex type definitions", () => {
     describe("A valid number with a specific maximum", () => {
         it("should validate", () => {
             expect(validateComplex(5, {
-                scalar: "number",
                 number: {
                     max: 5
                 }
@@ -250,7 +238,6 @@ describe("Complex type definitions", () => {
     describe("An invalid number with a specific maximum", () => {
         it("should not validate", () => {
             expect(validateComplex(6, {
-                scalar: "number",
                 number: {
                     max: 5
                 }
@@ -264,14 +251,12 @@ describe("Complex type definitions", () => {
 
         it("should validate a valid value", () => {
             expect(validateComplex(MY_NUMBER, {
-                scalar: "number",
                 number: { value: isMyNumber }
             })).toBeTruthy();
         });
 
         it("should not validate an invalid value", () => {
             expect(validateComplex(MY_NUMBER + 1, {
-                scalar: "number",
                 number: { value: isMyNumber }
             })).toBeFalsy();
         });
@@ -279,7 +264,6 @@ describe("Complex type definitions", () => {
 
     describe("When a number's isInteger option is specified", () => {
         const schema: ComplexSchema = {
-            scalar: "number",
             number: {
                 isInteger: true
             }
@@ -300,7 +284,6 @@ describe("Complex type definitions", () => {
 
     describe("When a floating point value is validated against a floating point constant", () => {
         const schema: ComplexSchema = {
-            scalar: "number",
             number: {
                 value: 10.1,
                 epsilon: 1.1
@@ -324,7 +307,7 @@ describe("Complex type definitions", () => {
         it("should throw", () => {
             expect(
                 () => validateComplex([], { })
-            ).toThrowError(/Schema only allows zero or one/);
+            ).toThrowError(/Schema only allows one/);
         });
     });
 });