+++ CHANGELOG.md
@@ -4,6 +4,6 @@
 
 * Renamed TypeDef -> Schema
 * Removed deprecated settings
-* Aggregation type is now inferred from the schema; no need to specify it directly
+* Type and aggregation form are now inferred from the schema; no need to specify directly
 * Simplified schema member naming: `arrayDef -> array, stringDef -> string` and so on
 * Added support for optional typing